Update the cytoscape alias to `cytoscape/src/cjs`
This updates the cytoscape alias with the recent changes to the
cytoscape build system.  When you pull this commit into your local copy,
make sure to `npm i cytoscape` to make sure your cytoscape dependency is
in-sync.

Note that when we make merges to master for releases, we should lock
down the commit that the cytoscape dependecy points to.  This would
ensure that prod builds are always the same.